+++ eslint-config/ts.js
@@ -1,12 +1,21 @@
-const tsProfile = {
-    parser: '@typescript-eslint/parser', // Specifies the ESLint parser
-    extends: [
-        'eslint:recommended',
-        'plugin:@typescript-eslint/recommended',
-        'plugin:@typescript-eslint/recommended-requiring-type-checking',
-        'eslint-config-prettier',
-    ],
-    rules: require('./blui-rules'),
-};
+const js = require('@eslint/js');
+const tseslint = require('typescript-eslint');
+const prettierRecommended = require('eslint-plugin-prettier/recommended');
+const typescriptEslintParser = require('@typescript-eslint/parser');
+const bluiRules = require('./blui-rules.js');
 
-module.exports = tsProfile;
+module.exports = [
+    js.configs.recommended,
+    tseslint.configs.eslintRecommended,
+    ...tseslint.configs.recommendedTypeChecked,
+    ...tseslint.configs.stylisticTypeChecked,
+    prettierRecommended,
+    {
+        languageOptions: {
+            parser: typescriptEslintParser
+        },
+        rules: {
+            ...bluiRules
+        },
+    }
+];